requirements.txt
fastmcp
pydantic-ai
python-dotenv
